requirements.txt
google-search-results
google-generativeai
together
dotenv
streamlit